server/requirements.txt
sqlalchemy==1.3.6
psycopg2==2.8.3
Flask==1.0.3
Flask-Migrate==1.8.0
Flask-REST-JSONAPI==0.30.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Flask-Bootstrap==3.3.7.1
passlib==1.7.1
python-dotenv==0.10.3